requirements.txt
requests
tqdm
biopython